chore: add linux musl build target

+++ build/build-all-by-docker.sh
@@ -11,20 +11,25 @@ ehfs=/go/src/mjpclab.dev/ehfs
 rm -rf "$prefix/output/"
 
 buildByDocker() {
-  gover="$1"
+  local tag="$1"
   shift
-  docker pull golang:"$gover"
+  docker pull golang:"$tag"
 
   docker run \
     --rm \
     -v "$prefix":"$ehfs" \
     -e EX_UID="$(id -u)" \
     -e EX_GID="$(id -g)" \
-    golang:"$gover" \
-    /bin/bash -c '
-      sed -i -e "s;://[^/ ]*;://mirrors.aliyun.com;" /etc/apt/sources.list;
-      apt-get update;
-      apt-get install -yq git zip;
+    golang:"$tag" \
+    /bin/sh -c '
+      if [ -e /etc/apt/sources.list ]; then
+        sed -i -e "s;://[^/ ]*;://mirrors.aliyun.com;" /etc/apt/sources.list;
+        apt-get update;
+        apt-get install -yq git zip;
+      elif [ -e /etc/apk/repositories ]; then
+        sed -i "s;://[^/ ]*;://mirrors.aliyun.com;" /etc/apk/repositories
+        apk add bash git zip
+      fi
       git config --global safe.directory "*"
       /bin/bash '"$ehfs"'/build/build.sh "$@";
       chown -R $EX_UID:$EX_GID '"$ehfs"'/output;
@@ -34,6 +39,10 @@ buildByDocker() {
 }
 
 gover=latest
+buildByDocker "$gover" "${builds[@]}"
+
+gover=alpine
+builds=('linux amd64 -musl' 'linux amd64,v2 -musl' 'linux amd64,v3 -musl' 'linux arm64 -musl')
 buildByDocker "$gover" "${builds[@]}"
 
 #gover=1.20
